requirements.txt
Flask==2.3.3
SpeechRecognition==3.10.0
pyttsx3==2.90
gTTS==2.3.1
pyaudio==0.2.11
openai==1.3.7
python-dotenv==1.0.0